Added user registration
Created the forms on register page and updated the register function to crate new users.

currently getting a bug with the database again that i dont understand
RuntimeError: The current Flask app is not registered with this 'SQLAlchemy' instance. Did you forget to call 'init_app', or did you create multiple 'SQLAlchemy' instances?